strict cast

+++ src/session.cpp
@@ -110,9 +110,9 @@ void Session::ApplyActions(const vector<unique_ptr<Action>>& acts) {
 string Session::ActsToStr(const vector<unique_ptr<Action>>& acts) const {
   std::ostringstream s;
   s << "[";
-  for (int i = 0; i < acts.size(); i++) {
+  for (int i = 0; static_cast<size_t>(i) < acts.size(); i++) {
     s << acts[i]->ToString();
-    if (i < acts.size() - 1) s << ",";
+    if (static_cast<size_t>(i) < acts.size() - 1) s << ",";
   }
   s << "]";
   return move(s.str());
